refactor: flatten isSentFromAdmin module
