+++ src/commands/reputation/setReputation.test.ts
@@ -1,7 +1,7 @@
 import type { ChatInputCommandInteraction, GuildMember, User } from 'discord.js';
 import { beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
 import { mockDeep, mockReset } from 'vitest-mock-extended';
-import { isAdmin } from '../../utils/isSentFromAdmin';
+import { isAdmin } from '../../utils/permission';
 import { getOrCreateUser, updateRep } from './_helpers';
 import { setReputation } from './setReputation';
 
@@ -9,7 +9,7 @@ vi.mock('./_helpers');
 const mockCreateUpdateUser = vi.mocked(getOrCreateUser);
 const mockUpdateRep = vi.mocked(updateRep);
 
-vi.mock('../../utils/isSentFromAdmin');
+vi.mock('../../utils/permission');
 const mockIsSentFromAdmin = vi.mocked(isAdmin);
 
 const mockInteraction = mockDeep<ChatInputCommandInteraction>();
