+++ src/commands/reputation/setReputation.ts
@@ -1,6 +1,6 @@
 import { type ChatInputCommandInteraction, type GuildMember, SlashCommandSubcommandBuilder } from 'discord.js';
-import { isAdmin, isModerator } from '../../utils/isSentFromAdmin';
 import { logger } from '../../utils/logger';
+import { isAdmin, isModerator } from '../../utils/permission';
 import type { Subcommand } from '../builder';
 import { getOrCreateUser, updateRep } from './_helpers';
